# [OPTIMIZATION] Explicitly make xidmap shards as nil